test: Add more tests for hard delete task
